+++ src/routes/(v2)/v2/issue-358/+page.svelte
@@ -1,11 +1,16 @@
 <script lang="ts">
+	import { page } from '$app/stores';
 	import { superForm } from '$lib/index.js';
 	import SuperDebug from '$lib/index.js';
 
 	export let data;
 
+	// eslint-disable-next-line svelte/valid-compile
+	const resetForm = $page.url.searchParams.has('reset');
+
 	const { form, errors, message, enhance } = superForm(data.form, {
-		applyAction: false
+		invalidateAll: 'force',
+		resetForm
 	});
 </script>